# Swagger/documentation.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api animalerie",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "sandbox-server"
        }
    ],
    "paths": {
        "/achat": {
            "get": {
                "tags": [
                    "Achat"
                ],
                "summary": "List all achats",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all achats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "achat[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Achat"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/achat/{id}": {
            "get": {
                "tags": [
                    "Achat"
                ],
                "summary": "Retourne un achat",
                "description": "Retourne un achat",
                "operationId": "7d2adf260fec9d24681905a684befe2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'achat à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne un achat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Achat"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/achat/create": {
            "post": {
                "tags": [
                    "Achat"
                ],
                "summary": "Create achat",
                "description": "Créer un achat",
                "operationId": "b8547408e1eb7a673de548b3fd391557",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Achat à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_utilisateur": {
                                        "type": "integer",
                                        "example": 32
                                    },
                                    "id_produit": {
                                        "type": "integer",
                                        "example": 19
                                    },
                                    "montant": {
                                        "type": "float",
                                        "example": 49.99
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 34
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Achat enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/achat/update/{id}": {
            "put": {
                "tags": [
                    "Achat"
                ],
                "summary": "Update achat",
                "description": "Update achat in Db",
                "operationId": "e11810d2719ff35ad316cda86929fe94",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'achat à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Update achat",
                    "description": "Achat à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_utilisateur": {
                                        "type": "integer",
                                        "example": 32
                                    },
                                    "id_produit": {
                                        "type": "integer",
                                        "example": 19
                                    },
                                    "montant": {
                                        "type": "float",
                                        "example": 49.99
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 34
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Achat enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/achat/delete/{id}": {
            "delete": {
                "tags": [
                    "Achat"
                ],
                "summary": "Delete achat",
                "description": "Delete achat in Db",
                "operationId": "5c17574b63caf48205b693ab2e1153ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'achat à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Achat supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Achat": {
                "title": "Achat",
                "properties": {
                    "id": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "id_utilisateur": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "id_produit": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "montatnt": {
                        "nullable": false,
                        "type": "float"
                    },
                    "quantite": {
                        "nullable": false,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}